# api_app/docs/swagger.yaml
definitions:
...
  api_model.AddParkingSlotRequest:
    properties:
      price:
        type: number
      total_slot:
        type: integer
      type:
        type: integer
    type: object
  api_model.BaseResponseAPIModel:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api_model.ListCityResponseDto:
    properties:
      city_id:
        type: integer
      city_name:
        type: string
    type: object
  api_model.ListDistrictResponseDto:
    properties:
      city_id:
        type: integer
      city_name:
        type: string
      district_id:
        type: integer
      district_name:
        type: string
    type: object
  api_model.ListWardResponseDto:
    properties:
      city_id:
        type: integer
      city_name:
        type: string
      district_id:
        type: integer
      district_name:
        type: string
      ward_id:
        type: integer
      ward_name:
        type: string
    type: object
...
  api_model.UpdateParkingSlotRequest:
    properties:
      parking_slot_id:
        type: integer
      price:
        type: number
      total_slot:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/internal/v1/:parking_id/parking-slot/:parking_slot_id/update:
    post:
      consumes:
      - application/json
      description: UpdateParkingSlotController
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api_model.UpdateParkingSlotRequest'
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: integer
      - description: parking_slot_id
        in: path
        name: parking_slot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: UpdateParkingSlotController
      tags:
      - internal
  /api/internal/v1/:parking_id/parking-slot/add:
    post:
      consumes:
      - application/json
      description: Add Parking Slot
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api_model.AddParkingSlotRequest'
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Add Parking Slot
      tags:
      - internal
...
  /api/public/v1/location/cities:
    get:
      consumes:
      - application/json
      description: Get All City
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.ListCityResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get All City
      tags:
      - public
  /api/public/v1/location/cities/{city_id}/districts:
    get:
      consumes:
      - application/json
      description: Get All District By CityId
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: city_id
        in: path
        name: city_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.ListDistrictResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get All District By CityId
      tags:
      - public
  /api/public/v1/location/cities/{city_id}/districts/{district_id}/wards:
    get:
      consumes:
      - application/json
      description: Get All Ward By CityId and DistrictId
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: city_id
        in: path
        name: city_id
        required: true
        type: integer
      - description: district_id
        in: path
        name: district_Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.ListWardResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get All Ward By CityId and DistrictId
      tags:
      - public
swagger: "2.0"